requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.4
convertdate==2.4.0
cryptography==41.0.2
datadispatch==1.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.3.0
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
exchangelib==4.9.0
filelock==3.12.2
fonttools==4.40.0
fpdf2==2.7.4
furl==2.1.3
graphviz==0.13.2
h11==0.14.0
hijri-converter==2.3.1
holidays==0.21.13
htmldocx==0.0.6
hubspot-api-client==4.0.6
identify==2.5.26
idna==3.4
importlib-metadata==4.13.0
isodate==0.6.1
java-access-bridge-wrapper==0.14.1
jsonpath-ng==1.5.3
jsonschema==4.18.1
jsonschema-specifications==2023.6.1
korean-lunar-calendar==0.3.1
loguru==0.7.0
lxml==4.9.3
mss==6.1.0
netsuitesdk==1.24.0
nodeenv==1.8.0
notifiers==1.3.3
numpy==1.25.2
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.2
orderedmultidict==1.0.1
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pdfminer.six==20221105
pendulum==2.1.2
Pillow==9.5.0
platformdirs==3.9.1
ply==3.11
pre-commit==3.3.3
psutil==5.9.5
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
PyMeeus==0.5.12
pynput-robocorp-fork==5.0.0
pyotp==2.8.0
pypdf==3.12.1
pyperclip==1.8.2
PySocks==1.7.1
pyspnego==0.9.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-xlib==0.33
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
requests-file==1.5.1
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
robocorp-log==2.6.0
robocorp-storage==0.2.0
robocorp-tasks==2.1.3
robotframework==5.0.1
robotframework-pythonlibcore==4.2.0
robotframework-requests==0.9.5
robotframework-seleniumlibrary==6.1.0
robotframework-seleniumtestability==2.1.0
rpaframework==23.5.2
rpaframework-core==11.0.2
rpaframework-pdf==7.1.5
rpds-py==0.8.10
selenium==4.9.1
simple-salesforce==1.12.4
six==1.16.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stringcase==1.2.0
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
tweepy==3.10.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
virtualenv==20.24.2
webdriver-manager==3.8.6
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.16.0
